fix: move users, articles routes to seprate files
