+++ routes/users.js
@@ -1,0 +1,7 @@
+const express = require("express");
+const { getUserInfo } = require("../controllers/user");
+
+const userRouter = express.Router();
+userRouter.get("/users/me", getUserInfo);
+
+module.exports = userRouter;
